requirements.txt
boto3==1.17.79
botocore==1.20.79
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
Flask==2.0.1
Flask-MySQL==1.5.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
py3dns==3.2.1
pycparser==2.20
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.17.1
s3transfer==0.4.2
six==1.16.0
urllib3==1.26.4
validate-email==1.3
Werkzeug==2.0.1
requests==2.25.1